+++ VisInfProc/filtermanager.h
@@ -10,8 +10,8 @@
 #include <cuda_runtime.h>
 
 extern void cudaCombineFilters(int sx, int sy, int sz,
-                               double* gpuTemp, double* gpuSpatial,
-                               double* gpuCombined);
+                               float* gpuTemp, float* gpuSpatial,
+                               float* gpuCombined);
 
 class FilterManager
 {
@@ -34,7 +34,7 @@ public:
      * @param type
      * @return
      */
-    static Buffer2D constructSpatialFilter(FilterSettings& s, double orientation, enum SpatialFilter type);
+    static Buffer2D constructSpatialFilter(FilterSettings& s, float orientation, enum SpatialFilter type);
     /**
      * @brief combineFilters Combines a temporal 1D and a spatial 2D filter into a spatial temporal 3D filter
      * @param temporal
@@ -45,8 +45,8 @@ public:
 
 
 private:
-    static inline float gaussTemporal(double sigma, double mu, double t);
-    static inline float gaussSpatial(double sigma, double x, double y);
+    static inline float gaussTemporal(float sigma, float mu, float t);
+    static inline float gaussSpatial(float sigma, float x, float y);
 };
 
 #endif // FILTERMANAGER_H
